requirements.txt
faiss-cpu-noavx2==1.6.3
numpy
flask
flask-restful
